Add multilingual UI support
