+++ README_zh.md
@@ -382,6 +382,7 @@ mtd-subtitle-web \
 ```
 
 打开 `http://127.0.0.1:7860`，上传音频/视频文件，审阅解析出的字幕分段，然后下载 JSON/SRT/ASS；若 `PATH` 中存在 `ffmpeg` 与 `ffprobe`，还可压制生成 MP4。
+Web 界面支持简体中文和英文。首次使用时会跟随浏览器语言，也可通过页头的语言选择器即时切换并记住选择。
 
 批量处理：
