+++ moss_transcribe_diarize/app/static/locales/en.json
@@ -1,0 +1,95 @@
+{
+  "app.title": "MOSS Subtitle Studio",
+  "locale.label": "Interface language",
+  "runtime.checking": "Checking",
+  "runtime.available": "FFmpeg available",
+  "runtime.unavailable": "FFmpeg missing",
+  "runtime.apiUnavailable": "API unavailable",
+  "runtime.connectionHelp": "Could not connect to api/runtime. Make sure this page is served by mtd-subtitle-web.",
+  "jobs.title": "Jobs",
+  "jobs.count.one": "{count} job",
+  "jobs.count.other": "{count} jobs",
+  "jobs.empty": "No jobs yet",
+  "jobs.current": "Current job",
+  "actions.refresh": "Refresh",
+  "actions.importMedia": "Import media",
+  "actions.startTranscription": "Start transcription",
+  "actions.deleteJob": "Delete job",
+  "actions.newJob": "New job",
+  "actions.burnVideo": "Burn video",
+  "actions.burning": "Burning...",
+  "actions.ffmpegUnavailable": "FFmpeg unavailable",
+  "actions.saveChanges": "Save changes",
+  "actions.saving": "Saving...",
+  "actions.retrySave": "Retry save",
+  "actions.rerun": "Transcribe again",
+  "actions.startRerun": "Start rerun",
+  "sidebar.expand": "Expand job sidebar",
+  "sidebar.collapse": "Collapse job sidebar",
+  "fields.mediaFile": "Media file",
+  "inference.settings": "Inference settings",
+  "inference.defaultsHint": "Defaults usually work",
+  "inference.prompt": "Inference prompt",
+  "inference.outputTokens": "Output tokens",
+  "inference.maxContext": "Context limit",
+  "inference.decoding": "Decoding",
+  "inference.temperature": "Temperature",
+  "processing.transcribing": "Transcribing",
+  "processing.creating": "Creating job",
+  "processing.preparing": "Uploading media and preparing transcription",
+  "processing.failed": "Job failed",
+  "segments.start": "Start",
+  "segments.end": "End",
+  "segments.speaker": "Speaker",
+  "segments.subtitle": "Subtitle",
+  "save.saved": "Saved",
+  "save.dirty": "Unsaved changes",
+  "save.saving": "Saving...",
+  "save.failed": "Save failed",
+  "save.failedWithDetail": "Save failed: {detail}",
+  "speaker.names": "Speaker names",
+  "speaker.none": "No speakers",
+  "speaker.displayName": "Display name",
+  "style.fontSize": "Font size",
+  "style.marginV": "Bottom margin",
+  "style.speaker": "Speaker",
+  "style.colors": "Colors",
+  "options.show": "Show",
+  "options.hide": "Hide",
+  "options.bySpeaker": "By speaker",
+  "options.uniform": "Uniform",
+  "outputs.title": "Outputs",
+  "outputs.transcript": "Transcript",
+  "rerun.title": "Rerun transcription",
+  "rerun.source": "Source media: {name}",
+  "notice.truncatedShort": "The output may be truncated. Increase output tokens and transcribe again.",
+  "notice.truncatedLong": "The generated token limit was reached. Check the end of the subtitles or rerun with a higher output token limit.",
+  "usage.pending": "Generate up to {max} tokens",
+  "usage.generated": "Generated {generated}/{max} tokens{prompt}",
+  "usage.prompt": " · prompt {prompt}",
+  "errors.uploadFailed": "Upload failed",
+  "errors.burnFailed": "Video burn failed",
+  "errors.rerunFailed": "Rerun failed",
+  "errors.saveFailed": "Save failed",
+  "errors.job_not_found": "The job does not exist or was deleted.",
+  "errors.job_running": "A running job cannot be deleted.",
+  "errors.media_missing": "The media file is missing.",
+  "errors.invalid_decoding": "Decoding must be greedy or sample.",
+  "errors.invalid_max_length": "The context limit must be greater than zero.",
+  "errors.invalid_max_new_tokens": "Output tokens must be greater than zero.",
+  "errors.invalid_temperature": "Sampling temperature must be greater than zero.",
+  "errors.invalid_segments": "The subtitle payload is invalid.",
+  "errors.ffmpeg_unavailable": "ffmpeg or ffprobe is not available on PATH.",
+  "errors.subtitles_unavailable": "No subtitles are available for this job.",
+  "errors.file_not_ready": "The requested file is not ready.",
+  "status.queued": "Queued",
+  "status.loading_model": "Loading model",
+  "status.transcribing": "Transcribing",
+  "status.postprocessing": "Processing",
+  "status.waiting_review": "Ready for review",
+  "status.rendering": "Burning",
+  "status.done": "Done",
+  "status.failed": "Failed",
+  "status.cancelled": "Cancelled",
+  "status.idle": "Idle"
+}
