+++ moss_transcribe_diarize/app/static/locales/zh-CN.json
@@ -1,0 +1,95 @@
+{
+  "app.title": "MOSS 字幕工作台",
+  "locale.label": "界面语言",
+  "runtime.checking": "检测中",
+  "runtime.available": "FFmpeg 可用",
+  "runtime.unavailable": "FFmpeg 缺失",
+  "runtime.apiUnavailable": "API 连接失败",
+  "runtime.connectionHelp": "无法连接 api/runtime，请确认页面来自 mtd-subtitle-web 服务。",
+  "jobs.title": "任务",
+  "jobs.count.one": "{count} 个任务",
+  "jobs.count.other": "{count} 个任务",
+  "jobs.empty": "还没有任务",
+  "jobs.current": "当前任务",
+  "actions.refresh": "刷新",
+  "actions.importMedia": "导入媒体",
+  "actions.startTranscription": "开始转写",
+  "actions.deleteJob": "删除任务",
+  "actions.newJob": "新建任务",
+  "actions.burnVideo": "烧录视频",
+  "actions.burning": "烧录中...",
+  "actions.ffmpegUnavailable": "FFmpeg 不可用",
+  "actions.saveChanges": "保存修改",
+  "actions.saving": "保存中...",
+  "actions.retrySave": "重试保存",
+  "actions.rerun": "重新转写",
+  "actions.startRerun": "开始重跑",
+  "sidebar.expand": "展开任务栏",
+  "sidebar.collapse": "收起任务栏",
+  "fields.mediaFile": "媒体文件",
+  "inference.settings": "推理参数",
+  "inference.defaultsHint": "默认不用改",
+  "inference.prompt": "推理 Prompt",
+  "inference.outputTokens": "输出 tokens",
+  "inference.maxContext": "上下文上限",
+  "inference.decoding": "解码",
+  "inference.temperature": "温度",
+  "processing.transcribing": "转写中",
+  "processing.creating": "创建任务",
+  "processing.preparing": "上传媒体并准备转写",
+  "processing.failed": "任务失败",
+  "segments.start": "开始",
+  "segments.end": "结束",
+  "segments.speaker": "说话人",
+  "segments.subtitle": "字幕",
+  "save.saved": "已保存",
+  "save.dirty": "有未保存修改",
+  "save.saving": "正在保存...",
+  "save.failed": "保存失败",
+  "save.failedWithDetail": "保存失败：{detail}",
+  "speaker.names": "说话人名称",
+  "speaker.none": "暂无说话人",
+  "speaker.displayName": "显示名称",
+  "style.fontSize": "字号",
+  "style.marginV": "底边距",
+  "style.speaker": "说话人",
+  "style.colors": "颜色",
+  "options.show": "显示",
+  "options.hide": "隐藏",
+  "options.bySpeaker": "按说话人",
+  "options.uniform": "统一",
+  "outputs.title": "输出",
+  "outputs.transcript": "原文",
+  "rerun.title": "重跑转写",
+  "rerun.source": "来源媒体：{name}",
+  "notice.truncatedShort": "可能截断，建议提高输出 tokens 后重新转写。",
+  "notice.truncatedLong": "可能截断：生成 token 已达到上限，请检查字幕末尾或提高输出 tokens 后重跑。",
+  "usage.pending": "生成 tokens {max}",
+  "usage.generated": "生成 {generated}/{max} tokens{prompt}",
+  "usage.prompt": " · prompt {prompt}",
+  "errors.uploadFailed": "上传失败",
+  "errors.burnFailed": "烧录失败",
+  "errors.rerunFailed": "重跑失败",
+  "errors.saveFailed": "保存失败",
+  "errors.job_not_found": "任务不存在或已被删除。",
+  "errors.job_running": "任务运行期间不能删除。",
+  "errors.media_missing": "媒体文件不存在。",
+  "errors.invalid_decoding": "解码方式必须是 greedy 或 sample。",
+  "errors.invalid_max_length": "上下文上限必须大于 0。",
+  "errors.invalid_max_new_tokens": "输出 tokens 必须大于 0。",
+  "errors.invalid_temperature": "采样温度必须大于 0。",
+  "errors.invalid_segments": "字幕数据格式无效。",
+  "errors.ffmpeg_unavailable": "PATH 中未找到 ffmpeg 或 ffprobe。",
+  "errors.subtitles_unavailable": "该任务还没有可用字幕。",
+  "errors.file_not_ready": "请求的文件尚未生成。",
+  "status.queued": "排队中",
+  "status.loading_model": "加载模型",
+  "status.transcribing": "转写中",
+  "status.postprocessing": "处理中",
+  "status.waiting_review": "待校对",
+  "status.rendering": "烧录中",
+  "status.done": "已完成",
+  "status.failed": "失败",
+  "status.cancelled": "已取消",
+  "status.idle": "空闲"
+}
